+++ src/zxing/zxing/common/reedsolomon/ReedSolomonEncoder.cpp
@@ -51,7 +51,7 @@ void ReedSolomonEncoder::encode(std::vector<zxing::byte> &toEncode, int ecBytes)
     ArrayRef<int> infoCoefficients(dataBytes);
 
     //to-do optimize the following loop
-    for(size_t i=0; i< dataBytes; i++)
+    for(int i=0; i< dataBytes; i++)
         infoCoefficients[i] = toEncode[size_t(i)];
 
     Ref<GenericGFPoly> info(new GenericGFPoly(field_, infoCoefficients));
